criacao das views e ajustes no controller e model

+++ Http/Controllers/BannersController.php
@@ -11,6 +11,7 @@ use Gate;
 use App\Helpers\Functions;
 use App\Models\Menu;
 use App\Models\Language;
+use App\Models\Translation;
 use Modules\Banners\Models\Banner;
 
 use Modules\Banners\Http\Requests\BannerRequest;
@@ -53,9 +54,13 @@ class BannersController extends Controller
         $menu_id = $this->menu_id;
         $menu_icon = $this->menu_icon;        
         $menu_name = $this->menu_name;
+        $slug = $this->slug;        
+
         if(!is_null($menu_id)){
             $banners = $banner->where('menu_id', $menu_id)->orderBy('order', 'asc')->paginate(50);
-            return view('Banner::index', compact('banners', 'menu_icon', 'menu_name'));
+            $total = $banners->total();
+            $orders = \Functions::number_array($total);
+            return view('Banner::index', compact('banners', 'menu_icon', 'menu_name', 'slug', 'orders'));
         }else{
             abort(403, 'Página não encontrada');
         }
@@ -63,31 +68,53 @@ class BannersController extends Controller
 
     public function create()
     {
+        if( Gate::denies("manager_{$this->slug}") ) 
+            abort(403, 'Você não tem permissão para gerenciar esta página');
+
         $menu_id = $this->menu_id;
         $menu_icon = $this->menu_icon;
         $menu_name = $this->menu_name;
-        return view('Banner::create', compact('menu_id', 'menu_icon', 'menu_name'));
+        $slug = $this->slug;
+        $combine_filds = $this->combine_filds;
+        $target_list = Banner::TARGET;
+        return view('Banner::create', compact('menu_id', 'menu_icon', 'menu_name', 'target_list', 'slug', 'combine_filds'));
     }
 
     public function store(BannerRequest $request)
     {
+        if( Gate::denies("manager_{$this->slug}") ) 
+            abort(403, 'Você não tem permissão para gerenciar esta página');
+
         $data = $request->only(array_keys($request->rules()));
+        if(isset($request->image))
+            $data['image'] = $this->_uploadImage($request);
         Banner::create($data);
         return redirect()->back()->with('success','Adicionado com sucesso!');
     }
 
     public function edit(Banner $banner)
     {
+        if( Gate::denies("manager_{$this->slug}") ) 
+            abort(403, 'Você não tem permissão para gerenciar esta página');
+
         $menu_id = $this->menu_id;
         $menu_icon = $this->menu_icon;
         $menu_name = $this->menu_name;
+        $slug = $this->slug;
+        $combine_filds = $this->combine_filds;
+        $target_list = Banner::TARGET;
         $languages = Language::where('status', 'active')->orderBy('order', 'asc')->get();
-        return view('Banner::edit', compact('banner', 'languages', 'menu_id', 'menu_icon', 'menu_name'));
+        return view('Banner::edit', compact('banner', 'languages', 'menu_id', 'menu_icon', 'menu_name', 'target_list', 'slug', 'combine_filds'));
     }
 
     public function update(BannerRequest $request, Banner $banner)
     {
+        if( Gate::denies("manager_{$this->slug}") ) 
+            abort(403, 'Você não tem permissão para gerenciar esta página');
+
         $data = $request->only(array_keys($request->rules()));
+        if(isset($request->image))
+            $data['image'] = $this->_uploadImage($request, $banner->image);
         $banner->fill($data);
         $banner->save();
         return redirect()->back()->with('success','Atualizado com sucesso');
@@ -95,7 +122,11 @@ class BannersController extends Controller
 
     public function destroy(Banner $banner)
     {
-        $banner->delete();
+        if( Gate::denies("manager_{$this->slug}") ) 
+            abort(403, 'Você não tem permissão para gerenciar esta página');
+            
+        Translation::where('parent_id', $banner->id)->where('menu_id', $this->menu_id)->delete();
+        $banner->delete();              
         return redirect()->back()->with('success','Excluído com sucesso!');
     }
 
@@ -147,6 +178,19 @@ class BannersController extends Controller
         }
         return redirect()->back()->with('success','Ordem atualizada com sucesso');
 
+    }
+
+    protected function _uploadImage(Request $request, $nameImage = null)
+    {
+        if(isset($request->image)){           
+            $responseUpload = \Upload::imagePublic($request, 'image', $this->folder, null, $nameImage);
+            if($responseUpload->original['success']){
+                return $responseUpload->original['file'];
+            }
+            return null;
+        }else{
+            return null;
+        }
     }
 
 }
